Document context/types refactor - provide document context for thumbnails
