+++ src/components/document/document.tsx
@@ -7,7 +7,8 @@ import { DocumentFileMenu } from "./document-file-menu";
 import { MyWorkDocumentOrBrowser } from "./document-or-browser";
 import { BaseComponent, IBaseProps } from "../base";
 import { ToolbarConfig } from "../toolbar";
-import { DocumentModelType, LearningLogDocument, LearningLogPublication } from "../../models/document/document";
+import { DocumentModelType } from "../../models/document/document";
+import { LearningLogDocument, LearningLogPublication } from "../../models/document/document-types";
 import { SupportType, TeacherSupportModelType, AudienceEnum } from "../../models/stores/supports";
 import { WorkspaceModelType } from "../../models/stores/workspace";
 import { IconButton } from "../utilities/icon-button";
