+++ src/hooks/use-document-context.ts
@@ -1,21 +1,14 @@
 import { useEffect, useRef } from "react";
-import { IDocumentContext } from "../components/document/document-context";
-import { DocumentModelType, ISetProperties } from "../models/document/document";
+import { DocumentModelType, getDocumentContext } from "../models/document/document";
+import { IDocumentContext } from "../models/document/document-types";
 import { usePrevious } from "./use-previous";
 
 export function useDocumentContext(document: DocumentModelType) {
   const prevDocument = usePrevious(document);
-  const documentContext = useRef<IDocumentContext>();
+  const documentContext = useRef<IDocumentContext>(getDocumentContext(document));
   useEffect(() => {
     if (document !== prevDocument) {
-      documentContext.current = {
-        type: document.type,
-        key: document.key,
-        title: document.title,
-        originDoc: document.originDoc,
-        getProperty: (key: string) => document.properties.get(key),
-        setProperties: (properties: ISetProperties) => document.setProperties(properties)
-      };
+      documentContext.current = getDocumentContext(document);
     }
   }, [prevDocument, document]);
   return documentContext.current;
